define structure of metric sets
A metricset encompasses one or more metrics, which can span from
storage to performance. If we design them in a modular way and use
JobSet + replicated jobs to structure, this might be a good design to
easily deploy several different metrics for an application. I have
explicitly chosen to use the word metric over benchmark because I
do not think we are storing benchmarks (for comparison) anywhere.
After this structural step I will work on designing the Kubernetes
operations that need to happen for each metric type, and further
develop the API definition/spec.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>