huge refactor of jobset structure
we were not meaningfully using jobset before. This refactors
so that the standalone metric set can be free to create whatever
structure of jobset is required, including a launcher/worker
design that is able to complete when the launcher is done. The
application and storage metric sets I have maintained to be simple
and create one JobSet->replicatedJob that either is monitoring a volume
or application of interest. I have made this choice because I think
it is the most logical / common use case, but of course can change
my mind / design if/when afforded.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>