add first draft of python sdk and logs parsing
Since we cannot control log output format, and I think requiring a database service adds
huge complexity, for this first effort I think we should simply provide structure
to sections of logs, and then have a custom Python module that can parse it on
behalf of the user. I am also adding the Python SDK to do that, and I have decided
to not automate this generation, but rather to make it simple for the user
to provide the YAML to generate (and then it is parsed depending on the metrics
discovered there.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>