add support for amg and interactive mode
AMG is one of the CORAL-2 benchmarks and the timings can provide
insights to performance. In addition, this changeset adds the test
for it (metrics.yaml manually applied), a Python example and parser
for a round of runs, an interactive mode for any metric to allow for
easier debugging, and the metric client delete command waits for pods
to terminate and disappear.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>