refactor to add common design jobset package
If this works it could be really cool - we have a jobs package that defines
common jobset (metric set) interfaces that can be shared between standalone
apps. As an example, the launcher/worker design works for both network
and traditional apps, with slight differences that are determined based
on the struct that inherits the interface. I hope this works!

Signed-off-by: vsoch <vsoch@users.noreply.github.com>