Refactor/standalone design (#41)
* add quicksilver
* save state of refactored app/network

This adds a shared template used for networking and app.
I think I can simplify this further to have a common shared
template.

* refactor to add common design jobset package

If this works it could be really cool - we have a jobs package that defines
common jobset (metric set) interfaces that can be shared between standalone
apps. As an example, the launcher/worker design works for both network
and traditional apps, with slight differences that are determined based
on the struct that inherits the interface. I hope this works!

* add pennant
* add ensure default name to success jobs
* simplify entrypoint script logic further

the common prefix to generate the hosts lists can be shared between
applications. We can also define it to be a more general prefix
field because it does not necessarily need to be mpirun

* quicksilver can use common prefix too
* also simplify storage interfaces!

This is so satisfying!

Signed-off-by: vsoch <vsoch@users.noreply.github.com>