add new design for perf metric
in this application design, we create an empty volume to
be shared by two containers in a pod. One container (with
a modular install of software) can move it along with
any other needed assets to the shared volume, and then
customize the application entrypoint to move the volume
assets where they might need to be and wait for files
to be present.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>