custom: add support for custom container
We should be able to support custom containers, and
configuration of addons to them. I am not liking the
design to have addons defined in parallel, and want to
refactor so they are part of the metric. I am also
wondering if the metrics themselves are more akin to
apps. I have not looked at this project in a bit and
need to think about it.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>